+++ src/lib/utils/functions/index.ts
@@ -1,2 +1,3 @@
 export * from './numbers.js';
 export * from './permissions.js';
+export * from './strings.js';
