+++ src/lib/utils/functions/strings.ts
@@ -1,0 +1,30 @@
+import { diffWordsWithSpace } from 'diff';
+import * as DJS from 'discord.js';
+
+/**
+ * Capitalize a sentence
+ * @param str The string to capitalize
+ */
+export function toCapitalize(str: string) {
+	return str.charAt(0).toUpperCase() + str.slice(1);
+}
+
+/**
+ * Trim a string to a certain length
+ * @param str The string to trim
+ * @param max The max length of the string
+ */
+export function truncate(str: string, max: number) {
+	return str.length > max ? `${str.slice(0, max - 3)}...` : str;
+}
+
+/**
+ * Get the difference between the old and new content
+ * @param oldString The old content
+ * @param newString The new content
+ */
+export function getDiff(oldString: string, newString: string): string {
+	return diffWordsWithSpace(DJS.escapeMarkdown(oldString), DJS.escapeMarkdown(newString))
+		.map(result => (result.added ? `**${result.value}**` : result.removed ? `~~${result.value}~~` : result.value))
+		.join('');
+}
